requirements.txt
Django==4.1
djangorestframework==3.14.0
requests==2.28.2
